feat: Added event tracking in analytics.js, Added ids for event tracking, restructured files in folders
