+++ index.html
@@ -14,17 +14,19 @@
 </head>
 <body>
     <script defer src="https://cloud.umami.is/script.js" data-website-id="64ff51c8-15da-49f3-b671-681e77484dc8"></script>
-    <script src="src/script/menu/menuController.js"></script>
+    <script src="src/script/analytics/analytics.js"></script>
     <script type="module" src="src/script/scrollAnimation.js"></script>
-    <script src="src/script/mouseHighlightEffect.js"></script>
-    <script type="module" src="src/script/highlightButton.js"></script>
+    <script src="src/script/menu/menuController.js"></script>
+    <script src="src/script/components/mouseHoverHighlight.js"></script>
+    <script type="module" src="src/script/components/highlightButton.js"></script>
     <!--<script src="src/script/cursor.js"></script>-->
     <script src="src/script/competence/cardLoader.js"></script>
     <script type="module" src="src/script/competence/scrollController.js"></script>
     <script type="module" src="src/script/competence/modalController.js"></script>
     <script type="module" src="src/script/competence/modalLoader.js"></script>
-    <script type="module" src="src/script/projectLoader.js"></script>
-    <script type="module" src="src/script/projectModal.js"></script>
+    <script type="module" src="src/script/project/projectLoader.js"></script>
+    <script type="module" src="src/script/project/projectModal.js"></script>
+    
     <nav class="nav">
         <div class="nav-header">
             <button class="nav-header-button" href="">MENU</button>
@@ -83,12 +85,12 @@
                     </div>
                 </div>
                 <div class="nav-main-item-button">
-                    <a class="nav-main-item-button-item" target="_blank" href="https://github.com/DeltaGamingCH">
+                    <a class="nav-main-item-button-item" id="menu-github-button" target="_blank" href="https://github.com/DeltaGamingCH">
                         <svg width="32" height="32" viewBox="0 0 96 96" xmlns="http://www.w3.org/2000/svg">
                             <path fill-rule="evenodd" clip-rule="evenodd" d="M48.854 0C21.839 0 0 22 0 49.217c0 21.756 13.993 40.172 33.405 46.69 2.427.49 3.316-1.059 3.316-2.362 0-1.141-.08-5.052-.08-9.127-13.59 2.934-16.42-5.867-16.42-5.867-2.184-5.704-5.42-7.17-5.42-7.17-4.448-3.015.324-3.015.324-3.015 4.934.326 7.523 5.052 7.523 5.052 4.367 7.496 11.404 5.378 14.235 4.074.404-3.178 1.699-5.378 3.074-6.6-10.839-1.141-22.243-5.378-22.243-24.283 0-5.378 1.94-9.778 5.014-13.2-.485-1.222-2.184-6.275.486-13.038 0 0 4.125-1.304 13.426 5.052a46.97 46.97 0 0 1 12.214-1.63c4.125 0 8.33.571 12.213 1.63 9.302-6.356 13.427-5.052 13.427-5.052 2.67 6.763.97 11.816.485 13.038 3.155 3.422 5.015 7.822 5.015 13.2 0 18.905-11.404 23.06-22.324 24.283 1.78 1.548 3.316 4.481 3.316 9.126 0 6.6-.08 11.897-.08 13.526 0 1.304.89 2.853 3.316 2.364 19.412-6.52 33.405-24.935 33.405-46.691C97.707 22 75.788 0 48.854 0z" fill="#fff"/>
                         </svg>
                     </a>
-                    <a class="nav-main-item-button-item" target="_blank" href="mailto:nikitaemmanuel.zenger@ksb-sg.ch">
+                    <a class="nav-main-item-button-item" id="menu-contact-button" target="_blank" href="mailto:nikitaemmanuel.zenger@ksb-sg.ch">
                         <p class="nav-main-item-button-item-text">Reach out</p>
                         <svg class="nav-main-item-button-item-svg" width="16" height="16" viewBox="0 0 30 23" fill="none" xmlns="http://www.w3.org/2000/svg">
                             <path d="M7 13L1.04023 9.27514C0.285734 8.80358 0.476538 7.65547 1.343 7.4533L29 1M7 13L8 22M7 13L29 1M8 22L14.5 17M8 22L11.5 15.5M14.5 17L20.0222 19.9735C20.5449 20.2549 21.1962 20.0218 21.4215 19.4726L29 1M14.5 17L11.5 15.5M29 1L11.5 15.5" stroke="#F9FBFD" stroke-linejoin="round"/>
