+++ src/script/components/mouseHoverHighlight.js
