+++ src/script/project/svgLoader.js
